+++ Monolithic/src/main/java/com/fivemybab/ittabab/store/command/application/service/StoreMenuService.java
@@ -3,8 +3,8 @@ package com.fivemybab.ittabab.store.command.application.service;
 import com.fivemybab.ittabab.exception.NotFoundException;
 import com.fivemybab.ittabab.store.command.application.dto.CreateStoreMenuDto;
 import com.fivemybab.ittabab.store.command.application.dto.UpdateStoreMenuDto;
-import com.fivemybab.ittabab.store.command.application.repository.StoreMenuRepository;
-import com.fivemybab.ittabab.store.command.application.repository.StoreOrderMenuRepository;
+import com.fivemybab.ittabab.store.command.domain.repository.StoreMenuRepository;
+import com.fivemybab.ittabab.store.command.domain.repository.StoreOrderMenuRepository;
 import com.fivemybab.ittabab.store.command.domain.aggregate.StoreMenu;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
