+++ Monolithic/src/main/java/com/fivemybab/ittabab/store/command/application/service/StoreService.java
@@ -3,10 +3,10 @@ package com.fivemybab.ittabab.store.command.application.service;
 import com.fivemybab.ittabab.exception.NotFoundException;
 import com.fivemybab.ittabab.store.command.application.dto.CreateStoreDto;
 import com.fivemybab.ittabab.store.command.application.dto.UpdateStoreDto;
-import com.fivemybab.ittabab.store.command.application.repository.StoreFavoriteRepository;
-import com.fivemybab.ittabab.store.command.application.repository.StoreMenuRepository;
-import com.fivemybab.ittabab.store.command.application.repository.StoreRepository;
-import com.fivemybab.ittabab.store.command.application.repository.StoreReviewRepository;
+import com.fivemybab.ittabab.store.command.domain.repository.StoreFavoriteRepository;
+import com.fivemybab.ittabab.store.command.domain.repository.StoreMenuRepository;
+import com.fivemybab.ittabab.store.command.domain.repository.StoreRepository;
+import com.fivemybab.ittabab.store.command.domain.repository.StoreReviewRepository;
 import com.fivemybab.ittabab.store.command.domain.aggregate.Store;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
