+++ Monolithic/src/main/java/com/fivemybab/ittabab/store/command/domain/repository/StoreOrderMenuRepository.java
@@ -1,7 +1,9 @@
-package com.fivemybab.ittabab.store.command.application.repository;
+package com.fivemybab.ittabab.store.command.domain.repository;
 
 import com.fivemybab.ittabab.store.command.domain.aggregate.StoreOrderMenu;
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
 
 public interface StoreOrderMenuRepository extends JpaRepository<StoreOrderMenu, Long> {
