+++ Monolithic/src/main/java/com/fivemybab/ittabab/store/command/domain/repository/StoreReviewRepository.java
@@ -1,4 +1,4 @@
-package com.fivemybab.ittabab.store.command.application.repository;
+package com.fivemybab.ittabab.store.command.domain.repository;
 
 import com.fivemybab.ittabab.store.command.domain.aggregate.StoreReview;
 import org.springframework.data.jpa.repository.JpaRepository;
